+++ sku.0/sys.server/compiled/game/script/library/missions.java
@@ -151,7 +151,7 @@ public class missions extends script.base_script
         int missionsCompleted = getPlayerDailyCount(player);
         if (missionsCompleted > DAILY_MISSION_CASH_REWARD)
         {
-            divisor += (float)missionsCompleted;
+            divisor += missionsCompleted;
         }
         return divisor;
     }
